Fix missing metrics on Windows in Docker/Containerd collectors
