+++ pkg/collector/corechecks/containers/docker/check_metrics_extension.go
@@ -37,9 +37,20 @@ func (dn *dockerCustomMetricsExtension) Process(tags []string, container *worklo
 		return
 	}
 
-	if containerStats.Memory != nil && containerStats.Memory.UsageTotal != nil && containerStats.Memory.Limit != nil && *containerStats.Memory.Limit > 0 {
-		memoryPct := *containerStats.Memory.UsageTotal / *containerStats.Memory.Limit
-		dn.sender(dn.aggSender.Gauge, "docker.mem.in_use", &memoryPct, tags)
+	if containerStats.Memory != nil {
+		// Re-implement Docker check behaviour: PrivateWorkingSet is mapped to RSS
+		if containerStats.Memory.PrivateWorkingSet != nil {
+			dn.sender(dn.aggSender.Gauge, "docker.mem.rss", containerStats.Memory.PrivateWorkingSet, tags)
+		}
+
+		if containerStats.Memory.SwapLimit != nil {
+			dn.sender(dn.aggSender.Gauge, "docker.mem.sw_limit", containerStats.Memory.SwapLimit, tags)
+		}
+
+		if containerStats.Memory.UsageTotal != nil && containerStats.Memory.Limit != nil && *containerStats.Memory.Limit > 0 {
+			memoryPct := *containerStats.Memory.UsageTotal / *containerStats.Memory.Limit
+			dn.sender(dn.aggSender.Gauge, "docker.mem.in_use", &memoryPct, tags)
+		}
 	}
 
 	if containerStats.CPU != nil {
