+++ pkg/util/cgroups/stats.go
@@ -54,7 +54,7 @@ type MemoryStats struct {
 	MinThreshold      *uint64 // cgroupv2 only
 	LowThreshold      *uint64 // cgroupv1: mapped to soft_limit
 	HighThreshold     *uint64 // cgroupv2 only
-	SwapLimit         *uint64
+	SwapLimit         *uint64 // Memory+Swap (thus >= Limit)
 	SwapHighThreshold *uint64 // cgroupv2 only
 
 	PSISome PSIStats
