+++ pkg/util/containers/v2/metrics/system/collector_linux.go
@@ -187,6 +187,7 @@ func buildMemoryStats(cgs *cgroups.MemoryStats) *provider.ContainerMemStats {
 	convertField(cgs.RSS, &cs.RSS)
 	convertField(cgs.Cache, &cs.Cache)
 	convertField(cgs.Swap, &cs.Swap)
+	convertField(cgs.SwapLimit, &cs.SwapLimit)
 	convertField(cgs.OOMEvents, &cs.OOMEvents)
 
 	return cs
